refactor(base-gen): extract ConfigFilterMixin (Tier-1b Task 3)
Move the three concrete config-filter helpers
(_filter_extractor_config_by_service_config, _get_extractor_state_service_name,
_filter_config_by_stride) verbatim into base_generation/_config_filter.py as
ConfigFilterMixin, composed into BaseGenerationService ahead of ABC. The abstract
config loaders (_load_extractor_config, _load_generation_service_config,
_create_extractor) stay physically on the base so __abstractmethods__ remains the
exact 8-name frozenset. Annotation-only self stubs for the base-owned attributes
the helpers read; no behavior or subclass-contract change.
